+++ ProdInCena/src/prodincena/producao/ShowBanda.java
@@ -45,12 +45,12 @@ public class ShowBanda extends Producao {
     }
 
     @Override
-    public int getDuracao() {
+    public double getDuracao() {
         return duracao;
     }
 
     @Override
-    public void setDuracao(int duracao) {
+    public void setDuracao(double duracao) {
         this.duracao = duracao;
     }
